requirements/local.txt
# Local development dependencies go here
-r base.txt

factory_boy==2.4.1
django-picklefield==0.3.1
